requirements.txt
pynamodb==4.3.1
ulid-py==0.0.12
dynamodb-json==1.3
boto3 #no-deploy
botocore #no-deploy